refactor: move getOffersOfProvider
to /tasks/common/api in preparation to move request server-side
