+++ src/domains/tasks/common/api/index.ts
@@ -1,4 +1,5 @@
 import axios, { AxiosRequestConfig } from "axios";
+import { Offer } from "../types";
 
 export const getTasksOfCustomerQuery = (config?: AxiosRequestConfig) => {
   return axios.get(`${process.env.NEXT_PUBLIC_API_URL}api/v1/tasks`, {
@@ -14,4 +15,14 @@ export async function getTasksQuery(config?: AxiosRequestConfig) {
     ...config,
     params: { page: currentPage, limit: 6 },
   });
+}
+
+export const getOffersUrl = `${process.env.NEXT_PUBLIC_API_URL}api/v1/offers`;
+
+export async function getOffersOfProviderQuery() {
+  const response = await axios.get(getOffersUrl, {
+    params: { my_offers: true },
+  });
+
+  return response.data.data.offers as Array<Offer>;
 }
